+++ packages/modules/narduk-mapkit/src/client/index.ts
@@ -1,4 +1,5 @@
 export * from './mapkit.js'
 export * from './layers.js'
+export * from './layout.js'
 export * from './runtime.js'
 export * from './temporal.js'